+++ src/Application/Features/Teams/Queries/GetAllTeams/GetAllHeroesQuery.cs
@@ -1,7 +1,6 @@
 using AutoMapper.QueryableExtensions;
 using Microsoft.EntityFrameworkCore;
 using SSW.CleanArchitecture.Application.Common.Interfaces;
-using SSW.CleanArchitecture.Domain.Teams;
 
 namespace SSW.CleanArchitecture.Application.Features.Teams.Queries.GetAllTeams;
 
@@ -18,20 +17,5 @@ public sealed class GetAllTeamsQueryHandler(
         return await dbContext.Teams
             .ProjectTo<TeamDto>(mapper.ConfigurationProvider)
             .ToListAsync(cancellationToken);
-    }
-}
-
-public class TeamDto
-{
-    public Guid Id { get; init; }
-    public required string Name { get; init; }
-}
-
-public class GetAllTeamsMapping : Profile
-{
-    public GetAllTeamsMapping()
-    {
-        CreateMap<Team, TeamDto>()
-            .ForMember(d => d.Id, opt => opt.MapFrom(s => s.Id.Value));
     }
 }
