+++ src/Application/Features/Teams/Queries/GetAllTeams/GetAllTeamsMapping.cs
@@ -1,0 +1,12 @@
+using SSW.CleanArchitecture.Domain.Teams;
+
+namespace SSW.CleanArchitecture.Application.Features.Teams.Queries.GetAllTeams;
+
+public class GetAllTeamsMapping : Profile
+{
+    public GetAllTeamsMapping()
+    {
+        CreateMap<Team, TeamDto>()
+            .ForMember(d => d.Id, opt => opt.MapFrom(s => s.Id.Value));
+    }
+}
